requirements.txt
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0